+++ Peon/Source/Game/Entity/Forge.hpp
@@ -7,6 +7,6 @@ public:
     Forge(Game* game, const glm::vec2& position);
     virtual ~Forge();
 
-    virtual void Update(double deltaTime);
+    virtual void Update(float deltaTime);
     virtual void Render(grim::graphics::SpriteBatch& spriteBatch);
 };
