+++ Peon/Source/Game/Entity/Stockpile.hpp
@@ -10,7 +10,7 @@ public:
 
     Inventory* GetInventory();
 
-    virtual void Update(double deltaTime);
+    virtual void Update(float deltaTime);
     virtual void Render(grim::graphics::SpriteBatch& spriteBatch);
 
 private:
